requirements.txt
streamlit==1.40.2
autogen-agentchat==0.7.5
autogen-ext[openai]==0.7.5
python-dotenv==1.0.1

